+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## 1.0.0-alpha.1
 
-* TODO: Describe initial release.
+* The initial alpha version of the HTTP REST library.
+* Middlewares (Interceptors)
+* Request/Response body converters
+* Multipart requests with progress
+* Request/Response logger
+